[GENERIC] - Remove the ability to have generic metatables.
It is no longer possible to set a metatable on the type of
a value*. Only values with individual metatables are permitted.
Greatly reduces complexity in terms of what is supported.

This was mostly used by the string library to allow accessing
it by indexing strings. While it allowed for more succinct syntax
(such as str:byte()) it also made strings stand out as semi-objects
as compared to other primitive values. Just use the string library
directly instead of calling it as a set of methods to strings.

* Excluding tables and userdata which did not support this.