[BINCALL] - Add 'self' parameter for calls to binary metamethods.
The binary metamethods now have an extra parameter and are
called as follows: '(self, lhs, rhs)'. This gives a 'self' in
'Obj:foo(lhs, rhs)' for convenience and also makes it possible
to know which argument is which without using lists or closures.
It also clarifies which operand is the "owner" of the call.

Note that 'self' is the operand which the call originated from.

In other words the 'self' is identical to 'lhs' and/or
'rhs' and lets one determine how to handle the operation
by using the 'self' argument as a pre-defined variable.