[PROXY] - A new kind of value enabling better abstraction.
A proxy is an entirely different value. As a basic type it is
an object with an association between its identity, a contained
value and an individual metatable. Operations on the proxy result
in a mechanism where calls to the metamethods instead receive the
contained value. By doing so the contained value is "proxied".

All metamethods support this in the core or through the API.
Only (raw) equality is supported as an operation by a proxy
without its own metatable. Anything else has to be defined.

Proxies can describe the mechanism of values without limitations.
They can give any value a unique identity through their own. Also
they provide encapsulation through indirection by the envelopment.