[MISC] - New parameter to 'lua_load' and related functions.
Create an extra parameter 'env' to 'lua_load' for specifying
which environment to use (defaults to the global table). Add
this to any routines using 'lua_load' as well. This is an
alternative to setting it manually with lua_setupvalue, a debug
API, which is no longer required for setting the environment.