[MISC] - Make the state avoid using global memory.
Rework 'luaO_nilobject' so that it is no longer a global
variable but instead part of the global state. This means
that any addresses inside of the state are from the allocator
and therefore the state is better contained. It enables a state
to be more independent of context and possibly shared between
different instances of a library. That is possible if the state
does not depend on any context specific to a particular instance.

This adds to a former commit which avoided storing userdata
pointers acquired from static C pointers in the registry. Such
usage of the registry should be a practice to avoid entirely.

The state's 'version' field remains an issue as a global.