[MISC] - Change the API to use a string for version of state.
The state's version as reported by 'lua_version' is now a
string instead of a number. It includes the 'release' as well.
This gets rid of the last usage of global memory in the state.
Also remove the check in 'luaL_checkversion' for multiple Lua
instances since this is not an issue anymore (or shouldn't be).